histogram/requirements.txt
beautifulsoup4==4.8.0
Flask==1.1.1
pymongo==3.9.0
PyYAML==3.13
selenium==3.14
requests==2.22.0